Add basic gpio snippet
I'm not sure if this is exactly what we want to do but it is a potential
direction to go in.

A random gpio section I have found is:

  gpio0: gpio@50000000 {
    compatible = "nordic,nrf-gpio";
    gpio-controller;
    reg = < 0x50000000 0x200 0x50000500 0x300 >;
    #gpio-cells = < 0x2 >;
    label = "GPIO_0";
    status = "okay";
    port = < 0x0 >;
    phandle = < 0x4 >;
  };

I'm not sure how many of those entries we need or how to deal with the
gpio@50000000 thing at the top that feels like it would need to be
driven by external information rather than a static snippets file but
maybe that is ok.

I suspect it is also preferable to write these in a yaml file and
convert them over to json in a build step.

The yaml entry can look like:

  New GPIO:
    prefix: "gpio"
    body: |
      gpio$1: {
          compatible = "${2:nordic,nrf-gpio}";
          status = "${3|okay,disabled|}";
      };
    description: "Adds new GPIO port section"